Refactor web chat folders
