+++ apps/web/src/chat/composer/ChatDraftContext.tsx
@@ -1,6 +1,6 @@
 import { createContext, useContext, useEffect, useState, type ReactElement, type ReactNode } from "react";
-import { useAppData } from "../appData";
-import type { PendingAttachment } from "./FileAttachment";
+import { useAppData } from "../../appData";
+import type { PendingAttachment } from "../attachments/FileAttachment";
 import {
   createChatDraftContent,
   loadChatDraftWorkspaceState,
@@ -10,7 +10,7 @@ import {
   type ChatDraftContent,
   type StoredChatDraft,
 } from "./chatDraftStorage";
-import { useOptionalChatSession } from "./sessionController";
+import { useOptionalChatSession } from "../sessionController";
 
 export type ChatDraft = Readonly<{
   workspaceId: string | null;
