+++ apps/web/src/chat/composer/useChatDictationCapture.ts
@@ -1,10 +1,10 @@
 import { useEffect, useRef, useState, type MutableRefObject, type RefObject } from "react";
-import { transcribeChatAudio } from "../api";
+import { transcribeChatAudio } from "../../api";
 import {
   explainBrowserMediaPermissionError,
   queryBrowserPermissionState,
-} from "../access/browserAccess";
-import type { TranslationKey, TranslationValues } from "../i18n";
+} from "../../access/browserAccess";
+import type { TranslationKey, TranslationValues } from "../../i18n";
 import {
   insertDictationTranscriptIntoDraft,
   type ChatDictationState,
