+++ apps/web/src/chat/layout/ChatToggle.tsx
